requirements.txt
dialogflow==1.1.0
google-cloud-dialogflow==2.0.0
python-telegram-bot==13.0
requests==2.24.0
vk-api==11.9.1
environ~=1.0
telegram~=0.0.1
PyYAML~=5.3.1